requirements.txt
Scrapy==2.11.0
scrapy-splash==0.9.0
PyYAML==6.0.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.8
coverage==7.3.1
